+++ sat/cli/diag/parser.py
@@ -12,11 +12,12 @@ import sat.parsergroups
 def add_diag_subparser(subparsers):
 
     xname_opts = sat.parsergroups.create_xname_options()
+    redfish_opts = sat.parsergroups.create_redfish_options()
 
     diag_parser = subparsers.add_parser(
         'diag', help='Launch diagnostics for Rosetta switches.',
         description='Launch diagnostics for Rosetta switches.',
-        parents=[xname_opts])
+        parents=[xname_opts, redfish_opts])
 
     diag_parser.add_argument('-t', '--timeout', default=300, metavar='SECONDS',
                              help='Timeout for a diagnostic run on a switch')