+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Man page of ``sat diag`` revised to better cover interactive mode.
 
 ### Added
+- ``sat cablecheck`` supports shared ``--redfish-username`` option.
 - Ability to list drives in ``sat hwinv`` with ``--list-drives``.
 - New fields for drive counts and total drive capacity when listing and
   summarizing nodes in ``sat hwinv``.